[Container] Move provider test classes to correct namespace.
