+++ tests/Classes/Container/Provider/ProviderClass.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Tests\Classes\Support;
+namespace Valkyrja\Tests\Classes\Container\Provider;
 
 use Valkyrja\Container\Provider\Contract\ProviderContract;
